+++ src/Products/CPUtils/Extensions/utils.py
@@ -713,11 +713,17 @@ def configure_ckeditor(self, default=1, allusers=1, custom='', rmTiny=1, forceTe
                "'JustifyRight','JustifyBlock'],\n['Table','SpecialChar'],\n'/',\n['Styles','Format'],"
                "\n['Maximize', 'ShowBlocks', 'Source']\n]",
                'plonemeeting':
-               u"[\n['Cut','Copy','Paste','PasteText','PasteFromWord','-',"
-               "'Scayt'],\n['Undo','Redo','-','RemoveFormat'],\n['Bold','Italic','Underline','Strike'],\n"
-               "['NumberedList','BulletedList','-','Outdent','Indent','Blockquote'],\n['JustifyLeft','JustifyCenter',"
-               "'JustifyRight','JustifyBlock'],\n['Table','SpecialChar','Link','Unlink','Image'],\n'/',\n"
-               "['Styles','Format'],\n['Maximize','ShowBlocks', 'Source']\n]",
+               u"[\n"
+               u"['Cut','Copy','Paste','PasteText','PasteFromWord','-','Scayt'],\n"
+               u"['Undo','Redo','-','RemoveFormat'],\n"
+               u"['Bold','Italic','Underline','Strike'],\n"
+               u"['NumberedList','BulletedList','-','Outdent','Indent'],\n"
+               u"['JustifyLeft','JustifyCenter','JustifyRight','JustifyBlock','FontSize'],\n"
+               u"['Table','SpecialChar','Link','Unlink','Image'],\n"
+               u"'/',\n"
+               u"['Styles','Format','NbSpace','NbHyphen'],\n"
+               u"['Maximize','ShowBlocks','Source']\n"
+               u"]\n",
                'ged':
                u"[\n['Cut','Copy','Paste','PasteText','PasteFromWord','-',"
                "'Scayt'],\n['Undo','Redo','-','RemoveFormat'],\n['Bold','Italic','Underline','Strike'],\n"