+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 1.8 (unreleased)
 ----------------
 
+- configure_ckeditor : added buttons 'FontSize', 'NbSpace' and 'NbHyphen' and
+  removed button 'Blockquote', from the default CKeditor custom toolbar used
+  for 'plonemeeting'.
+  [gbastien]
 - resources_order : function to list resources and output order.
   [sgeulette]
 - configure_ckeditor : disable tinymce resources.